langchain-server/requirements.txt
# ai-fastapi/requirements.txt

fastapi
uvicorn[standard]
pydantic
python-dotenv
requests

# LangChain and NVIDIA Integration
langchain
langchain-nvidia-ai-endpoints

# Vector Database and Embeddings
chromadb-client
langchain-community
sentence-transformers
